# contracts/openapi/control-plane-v1.yaml
openapi: 3.1.0
info:
  title: Molejo Control Plane API
  version: 1.0.0
  description: Pre-alpha control-plane contract for the local beta workspace.
servers:
  - url: /
paths:
...
  /api/v1/workspaces:
    get:
      operationId: listWorkspaces
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Authorized workspaces
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Workspace' } }
                  nextCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createWorkspace
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/OptionalIdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HierarchyInput' }
      responses:
        '202':
          description: Workspace creation accepted
          content:
            application/json:
              schema:
                type: object
                required: [workspace, operation]
                properties:
                  workspace: { $ref: '#/components/schemas/Workspace' }
                  operation: { $ref: '#/components/schemas/Operation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workspaces/{workspaceId}:
    parameters: [{ $ref: '#/components/parameters/WorkspaceId' }]
    get:
      operationId: getWorkspace
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Authorized workspace
          content: { application/json: { schema: { $ref: '#/components/schemas/Workspace' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateWorkspace
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HierarchyInput' } } }
      responses:
        '200':
          description: Updated workspace
          content: { application/json: { schema: { $ref: '#/components/schemas/Workspace' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/workspaces/{workspaceId}/projects:
    parameters: [{ $ref: '#/components/parameters/WorkspaceId' }]
    get:
      operationId: listProjects
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/IncludeArchived' }
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Project' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createProject
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HierarchyInput' } } }
      responses:
        '201':
          description: Created project
          content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workspaces/{workspaceId}/projects/{projectId}:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
    get:
      operationId: getProject
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Project
          content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateProject
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HierarchyInput' } } }
      responses:
        '200':
          description: Updated project
          content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: archiveProject
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '204': { description: Project archived }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workspaces/{workspaceId}/projects/{projectId}/environments:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
    get:
      operationId: listEnvironments
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/IncludeArchived' }
      responses:
        '200':
          description: Environments
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Environment' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createEnvironment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HierarchyInput' } } }
      responses:
        '201':
          description: Created environment
          content: { application/json: { schema: { $ref: '#/components/schemas/Environment' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workspaces/{workspaceId}/projects/{projectId}/environments/{environmentId}:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
      - { $ref: '#/components/parameters/EnvironmentId' }
    get:
      operationId: getEnvironment
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Environment
          content: { application/json: { schema: { $ref: '#/components/schemas/Environment' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateEnvironment
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HierarchyInput' } } }
      responses:
        '200':
          description: Updated environment
          content: { application/json: { schema: { $ref: '#/components/schemas/Environment' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: archiveEnvironment
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '204': { description: Environment archived }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/workspaces/{workspaceId}/projects/{projectId}/apps:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
    get:
      operationId: listApps
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/IncludeArchived' }
      responses:
        '200':
          description: Apps
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/App' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createApp
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HierarchyInput' } } }
      responses:
        '201':
          description: Created app
          content: { application/json: { schema: { $ref: '#/components/schemas/App' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workspaces/{workspaceId}/projects/{projectId}/apps/{appId}:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
      - { $ref: '#/components/parameters/AppId' }
    get:
      operationId: getApp
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: App
          content: { application/json: { schema: { $ref: '#/components/schemas/App' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateApp
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HierarchyInput' } } }
      responses:
        '200':
          description: Updated app
          content: { application/json: { schema: { $ref: '#/components/schemas/App' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: archiveApp
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '204': { description: App archived }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fruto_session
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1, maxLength: 64 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IncludeArchived:
      name: includeArchived
      in: query
      schema: { type: boolean, default: false }
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, pattern: '^ws-[a-z2-7]{20}$' }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, pattern: '^prj-[a-z2-7]{20}$' }
    EnvironmentId:
      name: environmentId
      in: path
      required: true
      schema: { type: string, pattern: '^env-[a-z2-7]{20}$' }
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string, pattern: '^app-[a-z2-7]{20}$' }
...
    OptionalIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Required for Secret parameter mutations; ignored for PlainText mutations.
      schema: { type: string, minLength: 1, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    Unauthorized:
      description: Authentication required
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authenticated user cannot perform this action
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Version or idempotency conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
...
    Workspace:
      type: object
      required: [id, name, version, state, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^ws-[a-z2-7]{20}$' }
        name: { type: string, minLength: 1, maxLength: 80 }
        version: { type: integer, minimum: 1 }
        state: { type: string, enum: [Pending, Running, Ready, Failed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    HierarchyInput:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
    Project:
      type: object
      required: [id, name, version, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^prj-[a-z2-7]{20}$' }
        name: { type: string, minLength: 1, maxLength: 80 }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        archivedAt: { type: [string, 'null'], format: date-time }
    Environment:
      type: object
      required: [id, name, version, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^env-[a-z2-7]{20}$' }
        name: { type: string, minLength: 1, maxLength: 80 }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        archivedAt: { type: [string, 'null'], format: date-time }
    App:
      type: object
      required: [id, name, version, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^app-[a-z2-7]{20}$' }
        name: { type: string, minLength: 1, maxLength: 80 }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        archivedAt: { type: [string, 'null'], format: date-time }
...
    Operation:
      type: object
      required: [id, kind, status, desiredVersion, attempts]
      properties:
        id: { type: string, pattern: '^op-' }
        appEnvironmentId: { type: string, pattern: '^aev-[a-z2-7]{20}$' }
        deploymentId: { type: string, pattern: '^dpl-[a-z2-7]{20}$' }
        kind: { type: string, enum: [ApplyDeployment, DeleteAppEnvironment, EnsureWorkspace] }
        status: { type: string, enum: [Pending, Running, Succeeded, Failed] }
        desiredVersion: { type: integer, minimum: 1 }
        attempts: { type: integer, minimum: 0 }
        errorCode: { type: string }
        errorMessage: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }